Am creeat baza de date, modificat un pic entitatile, am implementat repository pentru entitati(cu exceptia lui update, facem altcandva astea). Am creeat fisierele pentru ORM cu hibernate, am testat cateva functionalitati si merge ok.

+++ proiect_colectiv/src/main/java/com/proiect_colectiv/service/IService.java
@@ -3,13 +3,13 @@ package com.proiect_colectiv.service;
 import com.proiect_colectiv.model.Entity;
 
 public interface IService<E extends Entity> {
-    E findOne(int id);
+    E findOne(Long id);
 
     Iterable<E> findAll();
 
     void save(E entity);
 
-    void delete(int id);
+    void delete(Long id);
 
     void update(E entity);
 }
